+++ main
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 # coding=UTF-8
 #
 # Copyright (C) 2015  Michell Stuttgart
@@ -19,8 +21,8 @@ import sys
 from os.path import abspath
 from PySide import QtGui, QtCore
 
-from lib.main_window_model import MainWindowModel
-from lib.main_window_view import MainWindowView
+from pynocchio.src.main_window_model import MainWindowModel
+from pynocchio.src.main_window_view import MainWindowView
 
 DATADIRS = (
         abspath('.'),
