+++ pynocchio/locale/pynocchio_pt_BR.qm
