+++ pynocchio/src/bookmark_database_manager.py
@@ -16,7 +16,7 @@
 # with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from peewee import OperationalError, IntegrityError
-from bookmark import Bookmark, BookmarkBaseModel, db
+from pynocchio.src.bookmark import Bookmark, BookmarkBaseModel, db
 
 
 class BookmarkManager(BookmarkBaseModel):
