+++ pynocchio/src/comic.py
