+++ pynocchio/src/compact_file_loader.py
