+++ pynocchio/src/go_to_page_dialog.py
@@ -16,7 +16,7 @@
 # with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from PySide import QtCore, QtGui
-from uic_files.ui_go_to_page_dialog import Ui_GoPageDialog
+from pynocchio.src.uic_files.ui_go_to_page_dialog import Ui_GoPageDialog
 
 
 class GoToDialog(QtGui.QDialog):
