+++ pynocchio/src/pynocchio_exception.py
