+++ pynocchio/src/uic_files/custom_widgets/qscroll_area_viewer.py
