+++ pynocchio/src/uic_files/custom_widgets/status_bar.py
