+++ pynocchio/src/uic_files/main_window_view_rc.py
