+++ pynocchio/src/uic_files/ui_about_dialog.py
