+++ pynocchio/src/uic_files/ui_bookmark_manager_dialog.py
