+++ pynocchio/src/uic_files/ui_main_window_view.py
