+++ pynocchio/src/utility.py
