+++ setup.py
@@ -4,7 +4,7 @@ from setuptools import setup
 setup(
     name='pynocchio',
     version='0.0.1',
-    packages=['test', 'src'],
+    packages=['test', 'pynocchio'],
     test_suite='test',
     url='https://github.com/mstuttgart/pynocchio-comic-reader',
     license='GPL v3',
