--- src/lib/about_dialog.py
@@ -1,26 +1,0 @@
-# coding=UTF-8
-#
-# Copyright (C) 2015  Michell Stuttgart
-
-# This program is free software: you can redistribute it and/or modify it
-# under the terms of the GNU General Public License as published by the Free
-# Software Foundation, either version 3 of the License, or (at your option)
-# any later version.
-
-# This program is distributed in the hope that it will be useful, but WITHOUT
-# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-# FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
-# more details.
-
-# You should have received a copy of the GNU General Public License along
-# with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-from PySide import QtGui
-from uic_files.ui_about_dialog import Ui_AboutDialog
-
-
-class AboutDialog(QtGui.QDialog):
-    def __init__(self, parent=None):
-        QtGui.QDialog.__init__(self, parent)
-        self.ui = Ui_AboutDialog()
-        self.ui.setupUi(self)
